DECIMAL types in SQL should always be considered as longs, not ints

+++ sql/src/main/java/com/vmware/ddlog/ParseLiterals.java
@@ -55,7 +55,7 @@ class ParseLiterals extends SqlBasicVisitor<DDlogRecord> {
             case BOOLEAN:
                 return new DDlogRecord(sqlLiteral.booleanValue());
             case DECIMAL:
-                return new DDlogRecord(sqlLiteral.intValue(false));
+                return new DDlogRecord(sqlLiteral.longValue(false));
             case CHAR:
                 try {
                     return new DDlogRecord(sqlLiteral.toValue());
